hotfix: refatoração do componente grid
