+++ src/components/Grid/ImagensGrid/ImagensGrid.css
@@ -9,7 +9,7 @@
 }
 .item-grid .img-grid {
   width: 100%;
-  /* height: 100%; */
+  height: 100%;
   object-fit: cover;
   border-radius: 5px;
 }
