+++ src/components/Sobre/ConteudoSobre/ConteudoSobre.js
@@ -1,8 +1,8 @@
 import "./ConteudoSobre.css";
 import Texto from "../../Texto/Texto";
 import { NavLink } from "react-router-dom";
-import ImagensSobre from "../ImagensSobre/ImsagensSobre";
 import TextosSobre from "../TextosSobre/TextosSobre";
+import GridPosts from "../../Grid/GridPosts/GridPosts";
 
 const ConteudoSobre = (props) => {
   return (
@@ -29,7 +29,7 @@ const ConteudoSobre = (props) => {
         <TextosSobre id={props.id} />
       </div>
       <div className={props.id === "1" ? "grid-sobre__nos" : "grid-sobre__mee"}>
-        <ImagensSobre fotos={props.fotos} />
+        < GridPosts fotos={props.fotos} urlFotos={props.urlFotos} />
       </div>
     </div>
   );
